Repurpose old elasticsearch code for Index
* change the way it's initialised, add a method to make the transition
  easier
* add readers for index_name and field_names to make our existing 
  tests work
* ignore the Rakefile as the code we would change is going away anyway
